renamed SasFileParser class to SasFileParserExternal. It is quite confusing since there is already one from parso under the same package name.
