+++ src/main/scala/com/github/saurfang/sas/mapred/SasRecordReader.scala
@@ -3,8 +3,8 @@ package com.github.saurfang.sas.mapred
 import java.io.IOException
 
 import com.epam.parso.SasFileProperties
-import com.epam.parso.impl.SasFileParser
-import com.epam.parso.impl.SasFileParser.SasFileConstants
+import com.epam.parso.impl.{SasFileParserExternal, SasFileParser}
+import com.epam.parso.impl.SasFileParserExternal.SasFileConstants
 import com.github.saurfang.sas.util.PrivateMethodExposer
 import org.apache.commons.io.input.CountingInputStream
 import org.apache.hadoop.conf.Configuration
@@ -26,7 +26,7 @@ class SasRecordReader() extends RecordReader[NullWritable, Array[Object]] {
   private var splitEnd: Long = 0L
   private var fileInputStream: FSDataInputStream = null
   private var countingInputStream: CountingInputStream = null
-  private var sasFileReader: SasFileParser = null
+  private var sasFileReader: SasFileParserExternal = null
   private var sasFileReaderPrivateExposer: PrivateMethodExposer = null
   private var recordCount: Int = 0
   private var lastPageBlockCounter: Int = 0
@@ -83,7 +83,7 @@ class SasRecordReader() extends RecordReader[NullWritable, Array[Object]] {
     fileInputStream = fs.open(file)
     countingInputStream = new CountingInputStream(fileInputStream)
     // open SAS file reader to read meta data
-    sasFileReader = new SasFileParser.ExternalBuilder().sasFileStream(countingInputStream).build() // new SasFileReader(countingInputStream)
+    sasFileReader = new SasFileParserExternal.ExternalBuilder().sasFileStream(countingInputStream).build() // new SasFileReader(countingInputStream)
     sasFileReaderPrivateExposer = PrivateMethodExposer(sasFileReader)
 
     log.info(sasFileProperties.toString)
