+++ src/main/scala/com/github/saurfang/sas/util/PrivateMethodExposer.scala
@@ -1,5 +1,7 @@
 package com.github.saurfang.sas.util
 
+import com.epam.parso.impl.{SasFileParserExternal}
+
 class PrivateMethodCaller(x: AnyRef, methodName: String) {
   def apply(_args: Any*): Any = {
     val args = _args.map(_.asInstanceOf[AnyRef])
@@ -14,7 +16,7 @@ class PrivateMethodCaller(x: AnyRef, methodName: String) {
 
 /**
  *
- * Use this to invoke private methods in [[com.epam.parso.impl.SasFileParser]] so we don't need to modify it
+ * Use this to invoke private methods in [[SasFileParserExternal]] so we don't need to modify it
  * Credits to https://gist.github.com/jorgeortiz85/908035
  * Usage:
  *  p(instance)('privateMethod)(arg1, arg2, arg3)
